[#19] Rename: 파일명 camelCase 최종 변경
