+++ src/App.tsx
@@ -1,30 +1,30 @@
 import './App.css'
 import styled, {createGlobalStyle} from 'styled-components'
 import {createBrowserRouter, RouterProvider} from 'react-router-dom'
-import Home from './routes/hhome.tsx'
+import Home from './routes/home.tsx'
 import {useEffect} from 'react'
-import MyLetters from "./routes/mmyLetters.tsx";
+import MyLetters from "./routes/myLetters.tsx";
 import {ToastContainer} from "react-toastify";
 import OAuth from './components/oauth/KakaoOauth.tsx'
-import SignUp from './routes/ssignUp.tsx'
-import Mywish from './routes/mmyWish.tsx'
-import Login from './routes/llogin.tsx'
-import GuestLetters from "./routes/gguestLetters.tsx";
-import WriteLetter from "./routes/wwriteLetter.tsx";
-import LetterSentConfirm from "./routes/lletterSentConfirm.tsx";
-import Mypage from './pages/mmyPage.tsx'
-import EditMypage from './pages/mmyPageEdit.tsx'
-import WishList from './routes/wwishList.tsx'
-import UserWishList from './routes/uuserWishList.tsx'
-import GiftAmount from './pages/price/ggiftAmount.tsx'
-import UserWishDetail from './pages/uuserWishDetail.tsx'
-import SendConfirm from './pages/price/ssendConfirm.tsx'
-import PaymentRequest from "./routes/ppaymentRequest.tsx";
-import PaymentRequestComplete from "./routes/ppaymentRequestComplete.tsx";
-import MyWishDetail from "./pages/mmyWishDetail.tsx";
-import MyWishModify from "./pages/mmyWishModify.tsx";
-import AddWish from "./routes/aaddWish.tsx";
-import AddWishConfirm from "./routes/aaddWishConfirm.tsx";
+import SignUp from './routes/signUp.tsx'
+import Mywish from './routes/myWish.tsx'
+import Login from './routes/login.tsx'
+import GuestLetters from "./routes/guestLetters.tsx";
+import WriteLetter from "./routes/writeLetter.tsx";
+import LetterSentConfirm from "./routes/letterSentConfirm.tsx";
+import Mypage from './pages/myPage.tsx'
+import EditMypage from './pages/myPageEdit.tsx'
+import WishList from './routes/wishList.tsx'
+import UserWishList from './routes/userWishList.tsx'
+import GiftAmount from './pages/price/giftAmount.tsx'
+import UserWishDetail from './pages/userWishDetail.tsx'
+import SendConfirm from './pages/price/sendConfirm.tsx'
+import PaymentRequest from "./routes/paymentRequest.tsx";
+import PaymentRequestComplete from "./routes/paymentRequestComplete.tsx";
+import MyWishDetail from "./pages/myWishDetail.tsx";
+import MyWishModify from "./pages/myWishModify.tsx";
+import AddWish from "./routes/addWish.tsx";
+import AddWishConfirm from "./routes/addWishConfirm.tsx";
 
 const GlobalStyles = createGlobalStyle`
   body {
