+++ src/pages/myWishDetail.tsx
