+++ src/pages/price/inputPrice.tsx
