+++ src/pages/signup/getInfo.tsx
