+++ src/routes/signUp.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react"
-import GetInfo from "../pages/signup/ggetInfo";
-import AccountInput from "../pages/signup/aaccountInput";
-import AccountCheck from "../pages/signup/aaccountCheck";
+import GetInfo from "../pages/signup/getInfo";
+import AccountInput from "../pages/signup/accountInput";
+import AccountCheck from "../pages/signup/accountCheck";
 
 export default function SignUp() {
     const [page, setPage] = useState('getinfo');
